Fix bug #73:
The grabber.py example has an error in its implementation.
Method __del__ has been used as a C destroy function, producing
an error at exit because it was calling an invalid object.

Now, instead of a __del__ we use a deinitialize function to be called
by closeEvent generated by the mainWindow.